refactor the stepper code for readability and quality
